refactor(orm): relocate OrmValidationError to src/orm/errors/
RALPH: task=issue-73, prd=issue-72

## What
- Relocate OrmValidationError class, OrmValidationErrorKind, and
  OrmValidationFailure types from src/orm/schema-validations.ts to
  new src/orm/errors/validation.ts with barrel at src/orm/errors/index.ts
- Update all in-tree import sites to use the new path
- Re-export from schema-validations.ts for backward compatibility

## Key decisions
- Kept re-export in schema-validations.ts so the public API surface
  through src/orm/index.ts is unchanged — no breaking change
- Added export * from './errors' to src/orm/index.ts for direct access
- EquippedError import stays in schema-validations.ts because inline
  tests reference it

## Files changed
- src/orm/errors/validation.ts (new): relocated class and types
- src/orm/errors/index.ts (new): barrel re-export
- src/orm/schema-validations.ts: removed definitions, added import
  and re-export from ./errors
- src/orm/index.ts: added export * from './errors'
- src/orm/filter.ts: import from ./errors instead of ./schema-validations
- src/orm/adapters/postgresql/query.ts: import from ../../errors

## Blockers/notes
- All 471 tests pass, typecheck clean
- instanceof OrmValidationError preserved — class identity unchanged
