+++ src/orm/index.ts
@@ -1,6 +1,7 @@
 export * from './adapter'
 export * from './adapters/base'
 export { OrmAdapter } from './orm-adapter'
+export * from './errors'
 export * from './fields'
 export * from './filter'
 export * from './query'
